chore: API에 토큰 받아 인증 처리 추가

+++ family-moments/src/main/java/com/spring/familymoments/domain/commentLove/CommentLoveController.java
@@ -3,6 +3,7 @@ package com.spring.familymoments.domain.commentLove;
 import com.spring.familymoments.config.BaseException;
 import com.spring.familymoments.config.BaseResponse;
 import com.spring.familymoments.domain.commentLove.model.CommentLoveReq;
+import com.spring.familymoments.domain.user.AuthService;
 import com.spring.familymoments.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -11,6 +12,8 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.NoSuchElementException;
 
+import static com.spring.familymoments.config.BaseResponseStatus.INVALID_JWT;
+import static com.spring.familymoments.config.BaseResponseStatus.INVALID_USER_JWT;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 @RestController
@@ -20,6 +23,8 @@ public class CommentLoveController {
 
     private final CommentLoveService commentLoveService;
 
+    private final AuthService authService;
+
     /**
      * 댓글 내 하트 생성
      * [POST] /commentloves
@@ -27,7 +32,14 @@ public class CommentLoveController {
      */
     @PostMapping("")
     public BaseResponse<String> createPostLove(@AuthenticationPrincipal User user,
-                                               @RequestBody CommentLoveReq commentLoveReq) throws BaseException {
+                                               @RequestBody CommentLoveReq commentLoveReq,
+                                               @RequestHeader("X-AUTH-TOKEN") String requestAccessToken) throws BaseException, IllegalAccessException {
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+        if(user == null) {
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
+        }
 
         try {
             commentLoveService.createLove(user, commentLoveReq);
@@ -46,7 +58,15 @@ public class CommentLoveController {
      */
     @DeleteMapping("")
     public BaseResponse<String> deletePostLove(@AuthenticationPrincipal User user,
-                                               @RequestBody CommentLoveReq commentLoveReq) throws BaseException {
+                                               @RequestBody CommentLoveReq commentLoveReq,
+                                               @RequestHeader("X-AUTH-TOKEN") String requestAccessToken) throws BaseException, IllegalAccessException {
+
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+        if(user == null) {
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
+        }
 
         try {
             commentLoveService.deleteLove(user, commentLoveReq);
